+++ r-package/R/read_disaster_risk_area.R
@@ -14,7 +14,6 @@
 #'  For spatial analysis and statistics users should set simplified = FALSE. Borders have been
 #'  simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.
 #' @param showProgress Logical. Defaults to (TRUE) display progress bar
-#' @param tp Argument deprecated. Please use argument 'simplified'
 #'
 #' @export
 #' @examples \donttest{
@@ -29,9 +28,6 @@
 #'
 
 read_disaster_risk_area <- function(year=2010, simplified=TRUE, showProgress=TRUE, tp){
-
-  # deprecated 'tp' argument
-  if (!missing("tp")){stop(" 'tp' argument deprecated. Please use argument 'simplified' TRUE or FALSE")}
 
   # Get metadata with data url addresses
   temp_meta <- select_metadata(geography="disaster_risk_area", year=year, simplified=simplified)